reduce duplicate _make_divisible code
