+++ flowvision/models/ghostnet.py
@@ -10,6 +10,7 @@ import oneflow.nn.functional as F
 
 from .registry import ModelCreator
 from .utils import load_state_dict_from_url
+from ..utils import _make_divisible
 
 
 __all__ = ["ghostnet"]
@@ -18,22 +19,6 @@ __all__ = ["ghostnet"]
 model_urls = {
     "ghostnet": "https://oneflow-public.oss-cn-beijing.aliyuncs.com/model_zoo/flowvision/classification/GhostNet/ghostnet.zip",
 }
-
-
-def _make_divisible(v, divisor, min_value=None):
-    """
-    This function is taken from the original tf repo.
-    It ensures that all layers have a channel number that is divisible by 8
-    It can be seen here:
-    https://github.com/tensorflow/models/blob/master/research/slim/nets/mobilenet/mobilenet.py
-    """
-    if min_value is None:
-        min_value = divisor
-    new_v = max(min_value, int(v + divisor / 2) // divisor * divisor)
-    # Make sure that round down does not go down by more than 10%.
-    if new_v < 0.9 * v:
-        new_v += divisor
-    return new_v
 
 
 def hard_sigmoid(x, inplace: bool = False):
