+++ flowvision/models/rexnet_lite.py
@@ -8,6 +8,7 @@ import oneflow.nn as nn
 
 from .utils import load_state_dict_from_url
 from .registry import ModelCreator
+from ..utils import _make_divisible
 
 
 model_urls = {
@@ -16,27 +17,6 @@ model_urls = {
     "rexnet_lite_1_5": None,
     "rexnet_lite_2_0": None,
 }
-
-
-def _make_divisible(channel_size, divisor=None, min_value=None):
-    """
-    This function is taken from the original tf repo.
-    It ensures that all layers have a channel number that is divisible by 8
-    It can be seen here:
-    https://github.com/tensorflow/models/blob/master/research/slim/nets/mobilenet/mobilenet.py
-    """
-    if not divisor:
-        return channel_size
-
-    if min_value is None:
-        min_value = divisor
-    new_channel_size = max(
-        min_value, int(channel_size + divisor / 2) // divisor * divisor
-    )
-    # Make sure that round down does not go down by more than 10%.
-    if new_channel_size < 0.9 * channel_size:
-        new_channel_size += divisor
-    return new_channel_size
 
 
 def _add_conv(
