+++ flowvision/utils/__init__.py
@@ -3,3 +3,4 @@ from .clip_grad import dispatch_clip_grad
 from .vision_helpers import make_grid, save_image
 from .metrics import AverageMeter, accuracy
 from .model_ema import ModelEmaV2
+from .channel_helpers import _make_divisible
